+++ include/loops/transform.h
@@ -102,7 +102,8 @@
         (68,simdOps::SELUDerivative) ,\
         (70,simdOps::Reverse) ,\
         (71,simdOps::Pooling2D) ,\
-        (72,simdOps::MatchCondition)
+        (72,simdOps::MatchCondition) ,\
+        (73,simdOps::ClipByValue)
